+++ packages/plugin-hardhat/src/defender/utils.ts
@@ -40,13 +40,13 @@ export async function getNetwork(hre: HardhatRuntimeEnvironment): Promise<Networ
 
     const forkedNetworks = await networkClient.listForkedNetworks();
     const forkedNetwork = forkedNetworks.find((n) => n.chainId === chainId);
-    if (forkedNetwork) {
+    if (forkedNetwork !== undefined) {
       return forkedNetwork.name;
     }
 
     const privateNetworks = await networkClient.listPrivateNetworks();
     const privateNetwork = privateNetworks.find((n) => n.chainId === chainId);
-    if (privateNetwork) {
+    if (privateNetwork !== undefined) {
       return privateNetwork.name;
     }