+++ packages/plugin-hardhat/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-- Support private networks and forked networks with Defender.
+- Support private networks and forked networks with Defender. ([#989](https://github.com/OpenZeppelin/openzeppelin-upgrades/pull/989))
 
 ## 3.0.4 (2024-02-27)